feat: view, parser 분리
